Basic Report Missing Functionality and Fix ReadMe
This is an event driven approach to mesh network maintenance. Clients can report missing hosts to the discovery service when a specific request fails. They can then choose to retry (a new host) or fail out at that point. Zool Service Gateway instances will health check the reported host and if the health check fails (or timesout) the host will be removed from zool and remote clients will be updated.